refactor: MVVM 로직 수정중
