+++ WeatherForecast/WeatherForecast/ViewModel/WeatherTableViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  WeatherTableViewModel.swift
+//  WeatherForecast
+//
+//  Created by KimJaeYoun on 2021/10/15.
+//
+
+import Foundation
+
+final class WeatherTableViewModel {
+    private let service = WeatherService()
+
+    var onCompleted: (() -> Void)?
+}
+
